requirements.txt
telethon==1.32.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
coverage==7.2.7